+++ src/open_refinery/__init__.py
@@ -122,7 +122,7 @@ from .users import (
     valid_role,
 )
 
-__version__ = "0.13.0"
+__version__ = "0.13.1"
 
 __all__ = [
     "Factory",